test_main.py
import unittest
from main import select_language

class TestSelectLanguage(unittest.TestCase):
    def test_select_english(self):
        # Test selecting English
        # Replace the speak() function with a mock or comment it out for testing
        # Replace the get_voice_input() function with a mock or provide user input
        self.assertEqual(select_language(), 'en')

    def test_select_spanish(self):
        # Test selecting Spanish
        # Replace the speak() function with a mock or comment it out for testing
        # Replace the get_voice_input() function with a mock or provide user input
        self.assertEqual(select_language(), 'es')

    def test_quit_program(self):
        # Test quitting the program
        # Replace the speak() function with a mock or comment it out for testing
        # Replace the get_voice_input() function with a mock or provide user input
        with self.assertRaises(SystemExit):
            select_language()

if __name__ == '__main__':
    unittest.main()